Added and implemented testing for List_t container.
Cleaned up List_t internal implementation. By default, this list is now
expected to home heterogeneous items, and we simply leave the resource
management up to the items themselves as they are added to the List_t.
This does mean there has to be separate functions for Reference-type and
non-Reference type inserts, but it ends up simplifying the logic of
resource management substantially.

Added new internal List_Node_t functions for deleting a Node without
releasing it's contents. Also added new function to update the value
of an existing List_Node_t.

Updated the library header to wrap each container in a separate #define
guard. This is to allow accessing only the desired containers when
working with the library, rather than polluting the namespace with
everything even when only one is needed. See the updated README.md for
more details on this. Also added documentation to public functions
from the libcontainer.h main header.

Updated README.md with more information about the library and project.

Added new Build_Tuning.mk file to hold the build-time options for the
library in a more convenient representation.

Updated .clang-format rules to allow 100 columns per line. Ran
clang-format over the source files with the new rules.