Created new Hashmap_t container and HashFunc_t function type.
src/hashmap/ contains the full implementation of the Hashmap, which
is based on an Array_t of List_t to store the collisions in a chained
manner. This also includes default Hash functions for int, long, double,
and char* Key types.

Added new LIBCONTAINER_ENABLE_HASHMAP macro and API sections to the
libcontainer.h header.

Added new LIBCONTAINER_HASHMAP_LOAD_FACTOR and
LIBCONTAINER_HASHMAP_DEFAULT_CAPACITY build-time tuning parameters to
the top-level Makefile.

Updated NewComponent.sh to add the LICENSE file (if available) as a
preamble at the top of all generated files.

Updated ReleaseFunc_t typedef to no longer hide the pointer behind a
typedef. Updated all references to this type in the codebase.