Added callback functionality to Array_t container.
The Array_t container now provides an interface for performing either a
simple or complex callback function on each of the items it contains.

Array_DoCallback() allows for a function to access the element itself,
while Array_DoCallbackArg() extends this to pass an additional void
pointer which may contain arbitrary extra arguments to the Callback.
This is expected to operate in a similar way to pthread_create(), where
this additional pointer contains generally a struct or array of valuable data to
provide to the called function.

The Array_t will assert that the Value* (first argument of the callback)
is NEVER NULL.