Added callback functionality to Hashmap_t container.
Added similar Hashmap_DoCallback() and Hashmap_DoCallbackArg() functions
to the Hashmap_t container.

These are very similar to those for the Array_t and List_t containers,
with the exception of the "Value" pointer passed in to the callback
is actually a void*[2], containing a pointer to the Key and the Value of
the Hashmap entry respectively.