Working on Binary_Tree_t public API.
I have gotten the Create, Insert, Search, and Release functionality
working for the Binary Tree in a proper manner. I've also gotten
the DoCallback() and DoCallbackArgs() "working", but not in the manner I
want long-term.

I've removed the step-wise traversal API for now, as I need to plan out
how that has to work much more before I try to work on it.

Tree Remove() and Pop() functionality is still a work in progress.