Added new String_t Container.
The new String_t container is a rich wrapper for C strings to help
provide a consistent, memory safe interface, and which does not entirely
rely on NUL-terminated strings for operation. This container is
generally based on the std::String from the C++ STL, and the API will
likely expand to address additional functionality as it arises.
This String_t container also comes with a testing suite to unit test the
current functions.

String_t API Documentation to come in next commit.

Updated Makefiles to add a new QUIET flag to fully silence even the
pretty-printed output. Also updating to better integrate external
libraries into the build process. This isn't needed for this project,
but this Makefile *shoud* stay quite generic.

Updated README file to indicate the new String_t type.

Added new Array_Replace() and Array_ReplaceN() functions to allow
overwriting subslices of an Array with new contents in a safe manner,
without having to perform an inefficient RemoveN() and InsertN().

Updated all malloc() calls to instead use calloc() for better
consistency in structure initialization during the various *_Create()
functions.