Adding new API to query library version.
Added new version.c file and public API to get the version string,
version code, and compilation date of the library. These values are all
set by values defined in the top-level Makefile, and passed in as
compile-time defines. Added these version info functions as part of the
standard output when running the test target. Custom built the Makefile
for this version.c file to assert that it ALWAYS recompiles when the
library recompiles. This enforces that the version and compilation time
values are always correct.

Added new warning-ignore to suppress warnings/errors about unused
command-line arguments to the compiler. This was leading to some
erroneous warnings when passing some compiler flags to the archive tool
and vice-versa, despite the arguments being ignored.

Updated library README file to include a short snippet about the
built-in testing suite. Also added a short section about future
containers and planned extensions to the current containers.

Updated public String_t API to include variadic functions that accept a
va_list struct, in addition to the pure variadic functions. This API
better enables encapsulating these functions by simplifying the process
of passing variadic arguments down the call stack. Also updated the
String_t API to return a NUL char on error for the various Get*()
functions. Clarified that after a String_Unwrap(), the String_t provided
is released and ownership of the C-string is passed to the caller.

Slight internal modifications to the Hashmap_Entry_t and
Binary_Tree_Node_t objects to have their contents be an anonymous union,
to simplify the process of casting to/from the void* generic
representation.

Fixed a substantial inefficiency in Hashmap_rehash(), where it was
rehashing ALL of the buckets, rather than only the ones with existing
elements in them. Now, we only rehash the items in the original buckets
because any that end up in the new buckets MUST have already been
rehashed. Also fixed the item counting to be correct after each
successful reinsertion, rather than at the very end of the function.

Implemented the VERBOSE_TEST_SUCCESS macro, to suppress the "Test
Successful" print-out if this compilation option is selected.