Added Hashmap_Keys() API to access the set of held Keys in a Hashmap_t.
This new API provides a way to meaningfully search through a Hashmap_t,
as before this one would need to KNOW a key before they could find an
item. This function provides a method for getting the set of known keys
to allow iteration over and ensure that items can be searched for within
the map.

Added testing functions to validate new Hashmap_Keys() and
Hashmap_KeysSorted() functions.

Update of the library version to reflect the API update.

Fixed uncaught String_GetBack() bug with a read off the end of the
allocated memory.