Added *_FOREACH() Macro API for iterating over containers.
The new container-specific *_FOREACH() macros allow for simpler and
consistent iteration over the various containers of the library, using
the expected idioms for start and end detection.

Added new shorthand LIBCONTAINER_ENABLE_ALL macro to enable all
containers, rather than explicitly specifying all containers before
the #include.

Updated the definition of the CompareFunc_t from int(x)(void*, void*) to
int(x)(void*, void*, size_t), to better match with safer comparison
functions like strncmp() and memcmp(). Updated the implementations
within the library to use this additional parameter, and updated the
tests to also account for this.

Updated how an Array_t grows internally to always be a factor-of-two,
rather than based on the original capacity, to try and better match how
internal allocators generally work.

Updated interal Array_Sort() quicksort implementation to use a random
pivot, rather than always picking the end of the sub-array.

Fixed bug in how a Hashmap_Entry_t determines whether it holds memory or
just a reference to external memory. It no longer attempts to release
memory if the ReleaseFunc is NULL, rather than simply relying on a
ValueSize of 0.

Fixed missing NULL ReleaseFunc check in List_Node_t Release function.

Fixed off-by-one error in String_grow(), in the case where the array
needs to grow the length equal to the capacity. This had a chance of
causing the NUL-terminator to be written outside of the allocation.

Minor formatting fixes and format-style updates.