Added new Set_t container for an Ordered Set.
Added implementation, documentation, and testing suite for the new Set_t
container. This container provides an ordered, unique-only item
container, similar to the C++ std::ordered_set container. Added the
public typedefs, macros, and function API to main library header.
Added component-level Set_t test suite to the top-level main_test.c
file.

Added public SET_FOREACH macro to provide idiomatic iteration over a
Set_t.

Fixed all of the mismatched format strings. There were a number of "%ld"
format strings for "unsigned long", which is incorrect. Updated all of
these mismatches to the correct "%lu".

Fixed nagging off-by-one error in HashFunc_String(), where it would
attempt to use the NUL-terminator during computation of the hash value.

Fixed false-positive "node is not child of parent" error message in
Binary_Tree_t when removing nodes. It is entirely expected that the node
to be removed has an invalid Parent pointer as it's being removed.
Simply explicitly set this to NULL to reflect this and address the error
message. Added a validation function to the test suite to assert that
every Binary_Tree_Node_t within a tree satisfies the requirement that
the chain of (Node == Node->Parent->Child) is always true for one of the
child pointers.

Updated a few tests to the updated *_FOREACH() API as a cleanup.