Added new Binary_Heap_t container.
Added implementation, documentation, and testing suite for new
Binary_Heap_t container. This implements the expected binary heap
interface, and is either a min-heap or a max-heap depending on the given
comparison function. Added *_FOREACH() macro for idiomatic iteration
over the container.

Updated the default included CompareFunc implementations to more clearly
state that they provide an "Ascending" ordering, rather than leaving
this ambiguous. Fixed all references to these functions throughout the
library testing suite.

Cleaned up and simplified the Array_Clear() and Array_Release()
implementations to utilize the *_FOREACH() API, and to better handle the
edge case of releasing or clearing an empty Array_t.

Fixed some previously uncaught off-by-one errors in the Array_t
*_RemoveN(), *_ReplaceN() and *_GetElement() functions. Fixed an
out-of-bounds read when popping the final element from an Array_t when
the Length is equal to the Capacity.

Added a short-path for Array_swapItems() when the swap indices are
equal.

Fixed potential memory leak in Hashmap_Entry_Create() where the memory
for the Key would leak if the allocation for the Value failed.

Updated Hashmap testing suite to account for the NUL-terminator in all
cases where it handles C-Strings.

Fixed memory leak in Binary_Tree_Pop(), where memory could leak during
the removal of a node with two children. The ReleaseFunc values were not
transferred when the Successor overwrites the target node, leading to
memory leaking for the Key and Value pointers when the tree is later
released.