Removed variadic String_t functions.
As part of moving from C99 to ANSI-C compliance, one of the internal
memory-safety functions used in the variadic String_t functions
intrinsically required the C99 function "vsnprintf()". The only way of
removing the dependency on this function is to remove these variadic
functions.

Also removed the associated testing functions for the variadic String_t
functions.